pyproject.toml
[project]
name = "dynamictooling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.64.0",
]
